requirements.txt
# conda activate moot_court
torch==2.6.0
transformers==4.46.3
tokenizers==0.20.3
einops
addict
easydict
psutil
# pip install flash-attn==2.7.3 --no-build-isolation

# FastAPI and backend dependencies
fastapi==0.125.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-magic==0.4.27

# Existing project dependencies
langchain==0.2.0
langchain-core==0.2.0
langchain-huggingface==0.0.9
huggingface-hub==0.21.0
openai==1.3.7
google-generativeai>=0.7.0
mineru==0.2.1
pdf2image==1.16.3
